requirements.txt
ib_async>=2.0.1
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
plotly>=5.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
